# openapi.yaml
openapi: 3.0.2
info:
  contact:
    email: info@pdxfixit.com
    name: Ben Sandberg
    url: https://github.com/pdxfixit/hostdb-server
  description: Metadata storage service
  title: HostDB
  version: 0.3.2
...
servers:
  - description: Primary production instance of HostDB
    url: '{protocol}://hostdb.pdxfixit.com'
    variables:
      protocol:
        default: https
        enum:
          - http
          - https
paths:
...
  /stats:
    get:
      operationId: getStats
      responses:
        '200':
          $ref: '#/components/responses/stats'
        '500':
          $ref: '#/components/responses/error'
      summary: Get statistics about HostDB data.
      tags:
        - admin
...
components:
...
  responses:
...
    error:
      content:
        application/json:
          schema:
            properties:
              error:
                description: Error message.
                type: string
              reason:
                description: Optional, can provide context or debug hints.
                type: string
            required:
              - error
            type: object
        text/html:
          schema:
            type: string
      description: HostDB error
...
    stats:
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/stats'
      description: Statistical information about HostDB.
...
  schemas:
...
    stats:
      description: HostDB statistical information
      properties:
        lastseen_collectors:
          properties:
            hostdb-collector-aws:
              example: 2020-04-28 20:43:37
              type: string
            hostdb-collector-oneview:
              example: 2020-04-01 08:00:19
              type: string
            hostdb-collector-ucs:
              example: 2020-04-01 02:42:14
              type: string
            hostdb-collector-vrops:
              example: 2020-04-21 20:08:48
              type: string
          type: object
        newest_record:
          example: 2020-05-02 20:09:26
          type: string
        oldest_record:
          example: 2020-03-03 19:15:05
          type: string
        total_records:
          example: 84631
          type: integer
      type: object
...
  securitySchemes:
    BasicAuth:
      type: http
      scheme: basic
...